Move event simulation code
